+++ src/main/java/cart/application/CartItemService.java
@@ -4,9 +4,9 @@ import cart.domain.cart.Cart;
 import cart.domain.cart.CartItem;
 import cart.domain.cart.Quantity;
 import cart.domain.member.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.request.CartItemQuantityUpdateRequest;
+import cart.dto.request.CartItemRequest;
+import cart.dto.request.CartItemResponse;
 import cart.repository.CartRepository;
 import cart.repository.ProductRepository;
 import java.util.List;
