+++ src/main/java/cart/application/ProductService.java
@@ -4,8 +4,8 @@ import cart.domain.product.ImageUrl;
 import cart.domain.product.Name;
 import cart.domain.product.Price;
 import cart.domain.product.Product;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.dto.request.ProductRequest;
+import cart.dto.response.ProductResponse;
 import cart.repository.ProductRepository;
 import java.util.List;
 import java.util.stream.Collectors;
