+++ src/main/java/cart/controller/ProductApiController.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.application.ProductService;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.dto.request.ProductRequest;
+import cart.dto.response.ProductResponse;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.HttpStatus;
