+++ src/main/java/cart/dto/request/MemberResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 import cart.domain.member.Member;
